refactor(nms): Consolidate FEGNetworkContext
